# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Our site
  version: 0.1.0
paths:
...
  /api/users/register:
    post:
      tags:
      - Users
      summary: Регистрация нового пользователя
      operationId: register_api_users_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/users/me:
    get:
      tags:
      - Users
      summary: Получить профиль текущего пользователя
      operationId: get_me_api_users_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - AUTHENTICATED
      description: '**Требуемые роли:** Любой авторизованный пользователь'
    delete:
      tags:
      - Users
      summary: Удалить профиль текущего пользователя
      operationId: delete_user_me_api_users_me_delete
      responses:
        '204':
          description: Successful Response
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - AUTHENTICATED
      description: '**Требуемые роли:** Любой авторизованный пользователь'
    patch:
      tags:
      - Users
      summary: Обновить профиль текущего пользователя
      operationId: update_profile_me_api_users_me_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - AUTHENTICATED
      description: '**Требуемые роли:** Любой авторизованный пользователь'
  /api/users/{user_id}:
    get:
      tags:
      - Users
      summary: Получить пользователя по ID (Admin)
      operationId: get_by_id_api_users__user_id__get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
        '403':
          description: Недостаточно прав доступа для выполнения данной операции.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - ADMIN
      description: '**Требуемые роли:** `ADMIN`'
    patch:
      tags:
      - Users
      summary: Обновить профиль пользователя по ID (Admin)
      operationId: update_profile_api_users__user_id__patch
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
        '403':
          description: Недостаточно прав доступа для выполнения данной операции.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - ADMIN
      description: '**Требуемые роли:** `ADMIN`'
    delete:
      tags:
      - Users
      summary: Удалить пользователя по ID (Admin)
      operationId: delete_user_api_users__user_id__delete
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
        '403':
          description: Недостаточно прав доступа для выполнения данной операции.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - ADMIN
      description: '**Требуемые роли:** `ADMIN`'
  /api/users/me/change-password:
    post:
      tags:
      - Users
      summary: Сменить пароль текущего пользователя
      operationId: change_password_me_api_users_me_change_password_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
        required: true
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - AUTHENTICATED
      description: '**Требуемые роли:** Любой авторизованный пользователь'
  /api/users/{user_id}/change-password:
    post:
      tags:
      - Users
      summary: Сменить пароль пользователя по ID (Admin)
      operationId: change_password_api_users__user_id__change_password_post
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
        '403':
          description: Недостаточно прав доступа для выполнения данной операции.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - ADMIN
      description: '**Требуемые роли:** `ADMIN`'
...
components:
  schemas:
...
    ChangePasswordRequest:
      properties:
        old_password:
          type: string
          minLength: 1
          title: Old Password
        new_password:
          type: string
          maxLength: 128
          minLength: 6
          title: New Password
      type: object
      required:
      - old_password
      - new_password
      title: ChangePasswordRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RegisterUserRequest:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 2
          title: Name
          examples:
          - Иван Иванов
        email:
          type: string
          format: email
          title: Email
          examples:
          - ivan@edu.ru
        password:
          type: string
          maxLength: 128
          minLength: 6
          title: Password
          examples:
          - secretPassword123
        role:
          $ref: '#/components/schemas/UserRole'
          default: STUDENT
          examples:
          - STUDENT
      type: object
      required:
      - name
      - email
      - password
      title: RegisterUserRequest
...
    UpdateUserRequest:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 2
          - type: 'null'
          title: Name
        email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Email
      type: object
      title: UpdateUserRequest
...
    UserResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        email:
          type: string
          title: Email
        role:
          $ref: '#/components/schemas/UserRole'
      type: object
      required:
      - id
      - name
      - email
      - role
      title: UserResponse
    UserRole:
      type: string
      enum:
      - STUDENT
      - TEACHER
      - ADMIN
      title: UserRole
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    ErrorDetailResponse:
      title: ErrorDetailResponse
      type: object
      properties:
        detail:
          title: Detail
          type: string
          description: Описание ошибки
      required:
      - detail
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT Access Token (в заголовке `Authorization: Bearer <token>`
        или cookie `access_token`)'